Fix atomic file stores
os.link() fails if the target exists. The <name>.part target can
exist both because it's a simple, predictable name and because
we may have created it ourselves in a previous run that crashed.

Instead of <name>.part, use a fresh new temporary name every time.
For good measure, do this for the write() case, too.

Also make sure to delete the temporaries in case of any errors.

Perhaps this should probably apply to manifests etc. too, but this
commit does not attempt to fix them 'cause they ain't that broken.